+++ lib/Actions/Dashboard.php
@@ -8,7 +8,7 @@
  * captured response.
  *
  * @package cwp7
- * @version 2.3.0
+ * @version 2.4.0
  * @author  Booysen Logistics <bradley@booysenlogistics.co.za>
  * @license MIT
  * @link    https://github.com/bradleygb/whmcs-cwp-module